refactor: move `client` to `client/v5`
This will help with adding PiHole V6 support
